requirements.txt
# Document Extraction Pipeline Dependencies

# LlamaIndex Cloud SDK
# Note: llama-cloud 1.x is required for LlamaCloud class
# This may conflict with llama-index packages that require 0.x
llama-cloud>=1.2.0

# LandingAI ADE SDK
landingai-ade>=0.1.0

# HTTP clients
requests>=2.31.0
aiohttp>=3.9.0
httpx>=0.25.0

# Data validation and models
pydantic>=2.0.0

# Google Gemini for handwriting
google-genai>=1.0.0

# Environment and configuration
python-dotenv>=1.0.0

# Evaluation and analysis
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.13.0

# Async utilities
asyncio-throttle>=1.0.2
nest-asyncio>=1.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
Pillow>=10.0.0
python-magic>=0.4.27
pymupdf>=1.24.0

# LangChain for Agent (v0.3.x)
langchain==0.3.14
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# UI
streamlit>=1.30.0
lxml>=5.0.0
html5lib>=1.1
